+++ AssetBundle.php
@@ -17,7 +17,7 @@ class AssetBundle extends \yii\web\AssetBundle
     /**
      * @inherit
      */
-    public $sourcePath = '@vendor/rmrevin/yii2-fontawesome/assets';
+    public $sourcePath = '@vendor/fortawesome/font-awesome';
 
     /**
      * @inherit