+++ component/Stack.php
@@ -19,11 +19,13 @@ class Stack
 {
 
     /**
+     * @deprecated
      * @var string
      */
     public static $defaultTag = 'span';
 
     /**
+     * @deprecated
      * @var string
      */
     private $tag;
@@ -58,7 +60,23 @@ class Stack
      */
     public function __toString()
     {
-        return $this->render();
+        $options = $this->options;
+
+        $tag = ArrayHelper::remove($options, 'tag', 'span');
+
+        $template = ArrayHelper::remove($options, 'template', '{back}{front}');
+
+        $icon_back = $this->icon_back instanceof Icon
+            ? $this->icon_back->addCssClass(FA::$cssPrefix . '-stack-2x')
+            : null;
+
+        $icon_front = $this->icon_front instanceof Icon
+            ? $this->icon_front->addCssClass(FA::$cssPrefix . '-stack-1x')
+            : null;
+
+        $content = str_replace(['{back}', '{front}'], [$icon_back, $icon_front], $template);
+
+        return Html::tag($tag, $content, $options);
     }
 
     /**
@@ -94,6 +112,7 @@ class Stack
     }
 
     /**
+     * @deprecated
      * Change html tag.
      * @param string $tag
      * @return static
@@ -103,10 +122,13 @@ class Stack
     {
         $this->tag = $tag;
 
+        $this->options['tag'] = $tag;
+
         return $this;
     }
 
     /**
+     * @deprecated
      * @param string|null $tag
      * @param array $options
      * @return string