+++ tests/unit/fontawesome/MainTest.php
@@ -50,6 +50,11 @@ class MainTest extends \rmrevin\yii\fontawesome\tests\unit\TestCase
         );
 
         $this->assertEquals(
+            (string)FA::stack(['tag' => 'div']),
+            '<div class="fa-stack"></div>'
+        );
+
+        $this->assertEquals(
             (string)FA::stack()
                 ->icon('cog'),
             '<span class="fa-stack"><i class="fa fa-cog fa-stack-1x"></i></span>'
@@ -106,29 +111,36 @@ class MainTest extends \rmrevin\yii\fontawesome\tests\unit\TestCase
 
         $this->assertEquals(
             (string)FA::ul()
-                ->item('cog', 'Gear'),
-            '<ul class="fa-ul"><li><i class="fa fa-cog fa-li"></i>Gear</li></ul>'
+                ->item('Gear', ['icon' => 'cog']),
+            "<ul class=\"fa-ul\">\n<li><i class=\"fa fa-cog fa-li\"></i>Gear</li>\n</ul>"
         );
 
         $this->assertEquals(
             (string)FA::ul()
-                ->item('check', 'Check')
-                ->item('cog', 'Gear'),
-            '<ul class="fa-ul"><li><i class="fa fa-check fa-li"></i>Check</li><li><i class="fa fa-cog fa-li"></i>Gear</li></ul>'
+                ->item('Check', ['icon' => 'check'])
+                ->item('Gear', ['icon' => 'cog']),
+            "<ul class=\"fa-ul\">\n<li><i class=\"fa fa-check fa-li\"></i>Check</li>\n<li><i class=\"fa fa-cog fa-li\"></i>Gear</li>\n</ul>"
         );
 
         $this->assertEquals(
             (string)FA::ul()
                 ->tag('ol')
-                ->item('check', 'Check')
-                ->item('cog', 'Gear'),
-            '<ol class="fa-ul"><li><i class="fa fa-check fa-li"></i>Check</li><li><i class="fa fa-cog fa-li"></i>Gear</li></ol>'
+                ->item('Check', ['icon' => 'check'])
+                ->item('Gear', ['icon' => 'cog']),
+            "<ol class=\"fa-ul\">\n<li><i class=\"fa fa-check fa-li\"></i>Check</li>\n<li><i class=\"fa fa-cog fa-li\"></i>Gear</li>\n</ol>"
+        );
+
+        $this->assertEquals(
+            (string)FA::ul(['tag' => 'ol'])
+                ->item('Check', ['icon' => 'check'])
+                ->item('Gear', ['icon' => 'cog']),
+            "<ol class=\"fa-ul\">\n<li><i class=\"fa fa-check fa-li\"></i>Check</li>\n<li><i class=\"fa fa-cog fa-li\"></i>Gear</li>\n</ol>"
         );
 
         $this->assertEquals(
             (string)FA::ul()
-                ->item('check', 'Check', ['class' => 'another-class']),
-            '<ul class="fa-ul"><li class="another-class"><i class="fa fa-check fa-li"></i>Check</li></ul>'
+                ->item('Check', ['icon' => 'check', 'class' => 'another-class']),
+            "<ul class=\"fa-ul\">\n<li class=\"another-class\"><i class=\"fa fa-check fa-li\"></i>Check</li>\n</ul>"
         );
     }
 
@@ -140,6 +152,7 @@ class MainTest extends \rmrevin\yii\fontawesome\tests\unit\TestCase
 
         $this->assertEquals(FA::icon('cog'), '<i class="fontawesome fontawesome-cog"></i>');
         $this->assertEquals(FA::icon('cog')->tag('span'), '<span class="fontawesome fontawesome-cog"></span>');
+        $this->assertEquals(FA::icon('cog', ['tag' => 'span']), '<span class="fontawesome fontawesome-cog"></span>');
         $this->assertEquals(FA::icon('cog')->addCssClass('highlight'), '<i class="fontawesome fontawesome-cog highlight"></i>');
 
         $this->assertEquals(
@@ -151,8 +164,8 @@ class MainTest extends \rmrevin\yii\fontawesome\tests\unit\TestCase
 
         $this->assertEquals(
             (string)FA::ul()
-                ->item('cog', 'Gear'),
-            '<ul class="fontawesome-ul"><li><i class="fontawesome fontawesome-cog fontawesome-li"></i>Gear</li></ul>'
+                ->item('Gear', ['icon' => 'cog']),
+            "<ul class=\"fontawesome-ul\">\n<li><i class=\"fontawesome fontawesome-cog fontawesome-li\"></i>Gear</li>\n</ul>"
         );
 
         FA::$cssPrefix = $old_prefix;
@@ -163,19 +176,16 @@ class MainTest extends \rmrevin\yii\fontawesome\tests\unit\TestCase
         $this->assertEquals(FA::i('cog'), '<i class="fa fa-cog"></i>');
         $this->assertEquals(FA::icon('cog'), '<i class="fa fa-cog"></i>');
         $this->assertEquals(FA::icon('cog')->tag('span'), '<span class="fa fa-cog"></span>');
+        $this->assertEquals(FA::icon('cog', ['tag' => 'span']), '<span class="fa fa-cog"></span>');
         $this->assertEquals(FA::icon('cog')->addCssClass('highlight'), '<i class="fa fa-cog highlight"></i>');
 
         $this->assertEquals(FA::icon('cog')->inverse(), '<i class="fa fa-cog fa-inverse"></i>');
         $this->assertEquals(FA::icon('cog')->spin(), '<i class="fa fa-cog fa-spin"></i>');
         $this->assertEquals(FA::icon('cog')->fixedWidth(), '<i class="fa fa-cog fa-fw"></i>');
-        $this->assertEquals(FA::icon('cog')->fixed_width(), '<i class="fa fa-cog fa-fw"></i>');
-        $this->assertEquals(FA::icon('cog')->ul(), '<i class="fa fa-cog fa-ul"></i>');
         $this->assertEquals(FA::icon('cog')->li(), '<i class="fa fa-cog fa-li"></i>');
         $this->assertEquals(FA::icon('cog')->border(), '<i class="fa fa-cog fa-border"></i>');
         $this->assertEquals(FA::icon('cog')->pullLeft(), '<i class="fa fa-cog pull-left"></i>');
-        $this->assertEquals(FA::icon('cog')->pull_left(), '<i class="fa fa-cog pull-left"></i>');
         $this->assertEquals(FA::icon('cog')->pullRight(), '<i class="fa fa-cog pull-right"></i>');
-        $this->assertEquals(FA::icon('cog')->pull_right(), '<i class="fa fa-cog pull-right"></i>');
 
         $this->assertEquals(FA::icon('cog')->size(FA::SIZE_2X), '<i class="fa fa-cog fa-2x"></i>');
         $this->assertEquals(FA::icon('cog')->size(FA::SIZE_3X), '<i class="fa fa-cog fa-3x"></i>');
@@ -203,8 +213,8 @@ class MainTest extends \rmrevin\yii\fontawesome\tests\unit\TestCase
             'yii\base\InvalidConfigException',
             'FA::size() - invalid value. Use one of the constants: FA::SIZE_LARGE, FA::SIZE_2X, FA::SIZE_3X, FA::SIZE_4X, FA::SIZE_5X.'
         );
-        FA::icon('cog')
-            ->size('badvalue');
+
+        FA::icon('cog')->size('badvalue');
     }
 
     public function testIconRotateException()
@@ -213,8 +223,8 @@ class MainTest extends \rmrevin\yii\fontawesome\tests\unit\TestCase
             'yii\base\InvalidConfigException',
             'FA::rotate() - invalid value. Use one of the constants: FA::ROTATE_90, FA::ROTATE_180, FA::ROTATE_270.'
         );
-        FA::icon('cog')
-            ->rotate('badvalue');
+
+        FA::icon('cog')->rotate('badvalue');
     }
 
     public function testIconFlipException()
@@ -223,18 +233,20 @@ class MainTest extends \rmrevin\yii\fontawesome\tests\unit\TestCase
             'yii\base\InvalidConfigException',
             'FA::flip() - invalid value. Use one of the constants: FA::FLIP_HORIZONTAL, FA::FLIP_VERTICAL.'
         );
-        FA::icon('cog')
-            ->flip('badvalue');
+
+        FA::icon('cog')->flip('badvalue');
     }
 
     public function testIconAddCssClassCondition()
     {
+        $this->assertEquals(FA::$cssPrefix, 'fa');
         $this->assertEquals(FA::icon('cog')->addCssClass('highlight', true), '<i class="fa fa-cog highlight"></i>');
 
         $this->setExpectedException(
             'yii\base\InvalidConfigException',
             'Condition is false'
         );
+
         FA::icon('cog')->addCssClass('highlight', false, true);
     }
 }